refactor: 범위 체크 함수를 common 함수로 사용할 수 있도록 수정
