+++ src/view/InputView.js
@@ -1,5 +1,5 @@
 const { Console } = require('@woowacourse/mission-utils');
-const { isCollectBridgeLength, isValidateMoveInput, isValidateRetryInput } = require('../utils/validator.js');
+const { isCollectBridgeLength, isValidateMoveInput, isValidateRetryInput } = require('../utils/bridgeGameValidator.js');
 
 /**
  * 사용자로부터 입력을 받는 역할을 한다.
